pyproject.toml
[project]
name = "oscilloscope-vocabulary"
version = "0.1.0"
description = "Oscilloscope pattern visual vocabulary for harmonic-guided image generation"
readme = "README.md"
requires-python = ">=3.8"
authors = [{name = "Dal"}]
dependencies = [
    "fastmcp>=0.1.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/oscilloscope_vocabulary"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --tb=short"
